+++ packages/inqvine_core_main/lib/inqvine_core_main.dart
@@ -36,6 +36,7 @@ export 'package:flutter_riverpod/flutter_riverpod.dart';
 export 'package:hooks_riverpod/hooks_riverpod.dart';
 export 'package:riverpod/riverpod.dart';
 export 'package:logger/logger.dart';
+export 'package:beamer/beamer.dart';
 
 //* Imports